requirements.txt
tensorflow==2.10.0
scikit-learn==1.1.2
opencv-python==4.6.0.66
matplotlib==3.6.0
seaborn==0.12.0
numpy==1.23.3
pyyaml==6.0
fastapi==0.85.0
uvicorn==0.18.3
python-multipart==0.0.5
pytest==7.1.3
kaggle==1.5.12